LayoutTests/fast/forms/datalist/datalist-textinput-dynamically-add-options-expected.txt

PASS isShowingDataListSuggestions is false
PASS successfullyParsed is true

TEST COMPLETE
